# backend/artifacts/openapi/devsphere-backend-final.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DevSphere.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/applications/{jobApplicationId}/snapshot": {
      "get": {
        "tags": [
          "ApplicationHistory"
        ],
        "parameters": [
          {
            "name": "jobApplicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/applications/{jobApplicationId}/status-history": {
      "get": {
        "tags": [
          "ApplicationHistory"
        ],
        "parameters": [
          {
            "name": "jobApplicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter: Bearer {your JWT token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}